requirements.txt
# Core AI and orchestration
praisonai>=0.0.5
openai>=1.3.0
langchain>=0.1.0
pydantic>=2.0.0

# SIEM Integration
splunk-sdk>=1.7.0
requests>=2.31.0

# EDR Integration (CrowdStrike Falcon)
crowdstrike-falconpy>=1.3.0

# Malware Scanning
yara-python>=4.3.1

# Case Management (TheHive)
thehive4py>=1.13.0

# Data Processing and Analysis
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Logging and Monitoring
structlog>=23.1.0
rich>=13.0.0

# Environment and Configuration
python-dotenv>=1.0.0
click>=8.1.0

# HTTP and API
httpx>=0.24.0
aiohttp>=3.8.0

# Date/Time handling
python-dateutil>=2.8.0
pytz>=2023.3

# Testing dependencies removed

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Security
cryptography>=41.0.0 